Update README.md
Shell Redirection Project:
A script that prints Hello, World, followed by a new line to the standard output.
A script that displays a confused smiley: "(Ôo)'.
A script that displays the content of the /etc/passwd file.
A script that displays content of /etc/passwd and /etc/hosts.
A script that displays the last 10 lines of /etc/passwd.
A script that displays the third line of the file iacta.
A script that creates a file named exactly \*\\'"Best School"\'\\*$\?\*\*\*\*\*:)
A script that writes into the file ls_cwd_content the result of the command ls 	-la. 
A script that duplicates the last line of the file iacta.
A script that deletes all the regular files (not the directories) with a .js extension 
A script that deletes all the regular files (not the directories) with a .js extension 
A script that counts the number of directories and sub-directories in the current directory.
A Create a script that displays the 10 newest files in the current directory.
A script that takes a list of words as input and prints only words that appear exactly once.
A script that Displays lines containing the pattern “root” from the file /etc/passwd
A a script that Display the number of lines that contain the pattern “bin” in the file /etc/passwd
A script that Displays lines containing the pattern “root” and 3 lines after them in the file
A script that Displays all the lines in the file /etc/passwd that do not contain the pattern “bin”.
A script that Displays all lines of the file /etc/ssh/sshd_config starting with a letter. 
A script that Replaces all character’s A and c from input to Z and e respectively.
A script that removes all letters c and C from input.
A script that reverses its input.
A Write a script that displays all users and their home directories, sorted by users.
A script that finds all empty files and directories in the current directory and all sub-directories
A script that lists all the files with a .gif extension in the current directory and all its sub-directories.
A script that decodes acrostics that use the first letter of each line	

Feel free to go through my project, Feedback is highly Welcome